Move tests to a protocol-14/ subfolder
